Button to open Google
I've added an activity with a button to open the broswer, I've tried to connect it to the main screen to test it but I think we need to play around with the @Overide for the OnCreate events. I have also added code so that I can test the code on an Androis OS version 9 so I could put it on my phone. Everyone might need to go to the SDK manager and add Android version 9 to their android studio too.