Filter No longer crashes app
I've made changes to the filter and done A LOT of testing, but I need to do further testing to see why it does not work, I changed the functions to return a manipulated list but still nothing. Maybe we just need to refresh the page or something but I'll keep working on it.